requirements.txt
# Core framework - keep stable
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Auth - fix bcrypt issue only
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Utilities
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.1

# AI integration
openai==1.3.7

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Background tasks
celery==5.3.4
redis==5.0.2
flower==2.0.1